Improve protovalidate with managed mode
Update protovalidate to work with managed mode (without exceptions) and
'buf generate' as well as remote packages. Currently, those will
generate references to the protovalidate extensions to locations other
than the 'java_package' of 'build.buf.protovalidate' and when validation
is performed, an IllegalArgumentException is thrown.

Update ConstraintResolver to use getField instead of getExtension to
avoid the exception and transparently reparse extensions to the right
classes to be used by protovalidate-java.

Add a test exercising this behavior which generates code using managed
mode (without exceptions and '--include-imports') as well as one with
managed mode (with an exception).