Implement structured field and rule paths for violations (#213)
Implements the changes needed to pass conformance after
https://github.com/bufbuild/protovalidate/pull/265.

This is basically a straight port of what is done in
https://github.com/bufbuild/protovalidate-go/pull/154, owing to the
similarities between protovalidate-go and protovalidate-java.
Unfortunately this means it inherits some of the trickier maneuvers
needed to weave the right data into the right places. This version is
also not as efficient as a result of differences between Go and Java
protobufs, but effort was made to try to prevent big regressions in the
case of successful validation.

Some of this is probably not as pretty as it could be. I'm open to
improvements if there are any obvious things (I am certainly not a Java
expert after all) but if possible I'd like to defer anything that would
require major rethinking to down the road. (Happy to add TODOs for
those, though.)

DO NOT MERGE until the following is done:
- [x] https://github.com/bufbuild/protovalidate/pull/265 is merged
- [x] Dependencies updated to stable version release (waiting on manage
module push)