Add field and rule value to violations (#215)
Adds the ability to access the captured rule and field value from a
Violation.

There's a bit of refactoring toil, so I tried to split the commits
cleanly to make this easier to review.
- The first commit adds the `Violation` wrapper class and plumbs it
through all uses of `Violation`.
- The second commit makes `Value` non-internal so we can use it to
expose protobuf values in a cleaner fashion.
- The third commit actually implements filling the `fieldValue` and
`ruleValue`fields.

**This is a breaking change.** The API changes in the following ways:
- `ValidationResult` now provides a new wrapper `Violation` type instead
of the `buf.validate.Violation` message. This new wrapper has a
`getProto()` method to return a `buf.validate.Violation` message, and
`ValidationResult` now has a `toProto()` method to return a
`buf.validate.Violations` message.