Fix list binding (#276)
This fixes the way lists were being bound in the EvaluatorBuilder. The
call to
`DescriptorMappings.protoKindToCELType(fieldDescriptor.getType())` was
incorrect because the field descriptor type will be the type of the
elements in the list and not the list itself. This was resulting in a
list of ints getting bound to just an `int` value.