Fix CEL bindings for maps (#278)
This fixes the variable binding to `this` for maps. Previously, the
binding was using a map of the custom value abstraction `ObjectValue`
that we use in protovalidate-java. This `Value` class has no relation at
all to CEL-Java's `Val` object. So, using it as a variable binding does
not work because CEL-Java doesn't know how to interpret it.

This adds an additional `mapValueAsObject` function to return a map of
`Object`s for the binding. Note this not ideal, but should be sufficient
for now. Further context in the comments.