Fix operator in ValidationResult toString (#286)
Fixes #285 

This fixes the check for field path elements when calculating the
`toString` value for `ValidationResult` and adds a few tests.

Note that this also changes the value to print `Validation OK` if there
are no violations since it didn't make much sense to print `Validation
Error:` if there are no violations.

Technically this is a breaking change.