Add supplemental format conformance tests (#294)
The current conformance tests for `string.format` are not exhaustive and
do not account for all scenarios in the
[docs](https://github.com/google/cel-spec/blob/master/doc/extensions/strings.md).
One such example is a test for invalid UTF-8.

This adds the ability to specify supplemental conformance tests in the
form of another textproto file. The content is merged with the actual
cel conformance tests and then run against our implementation. This
allows us to specify our own tests not yet covered in the official
conformance tests. As a result, this includes two tests for invalid
UTF-8, which incidentally turned up a bug involving collapsing
placeholders for contiguous invalid UTF-8 bytes.

Note that a PR has been created
[here](https://github.com/google/cel-spec/pull/473) to add these tests
to the spec. Once added and released, they can be removed from our
supplemental tests.