Fix overload declaration for format (#295)
After some investigation, it seems the reason some of the cel
conformance tests weren't running was because the overload declaration
for `format` inside protovalidate-java was too narrow and was only
allowing primitive types inside lists.

This simplifies the declaration so that any `DynType` is accepted, which
allows maps as well as custom Protobuf types to be passed inside a list
to format.