Implement workaround for cel-java import issue (#302)
When collecting dependencies for types in cel-java, for some reason,
only public imports are considered. Because of this, if a message type
from another file is referenced within an expression, the type will not
be resolved, leading to strange, unexpected errors. In most cases,
things will work purely by accident, obscuring the broken resolution
behavior.

We should try to follow this up with an upstream fix as soon as
possible.