Reduce visibility of previously internal package
As part of refactoring, the previous internal package (where the
validator implementation is found) was refactored to the
build.buf.protovalidate package. As part of this change, we should
ensure that all previously internal classes are final (to prevent
subclassing) and that methods not overridden from an interface use
package private or lower visibility.